+++ changelog.md
@@ -1,4 +1,12 @@
-# <cc>v2.0.6</c>
+# <cc>v2.0.7</c>
+- <c-ffffc4><cy>**Updated**</c> Argon verification *further*.</c>
+- <c-ffffc4><cy>**Updated**</c> The Gauntlet Keeper's requirements from the Doom Gauntlet to just 10 Gauntlets total.</c>
+    - Their dialog updates with your Gauntlet progress!</c>
+- <c-ffffc4><cy>**Updated**</c> the Leaderboard to have placement numbers for the Top 100 players.</c>
+
+---
+
+## v2.0.6
 - <c-ffffc4><cy>**Updated**</c> Argon to now handle dead tokens accordingly.</c>
 
 ---